rearrange core lib
