+++ specs/STATE.md
@@ -2,9 +2,20 @@
 
 ## Last Updated
 
-2026-04-27 - Release version bumps for publish
+2026-05-25 - Move reusable portal adapters into packages/mcp-core
 
 ## Recent Changes
+
+### Move reusable portal adapters into packages/mcp-core (2026-05-25)
+
+- **Created `packages/mcp-core/src/backends/aomi-client.ts`** — moved `buildBackendPort` + `BackendPortDeps` from portal's `mcp-backend-bridge.ts`. Pure `BackendPort` implementation using `AomiClient`, zero portal-specific logic.
+- **Created `packages/mcp-core/src/adapters/auth-adapter.ts`** — extracted `buildAuthPort` + `AuthPortDeps` from portal's `mcp-server.ts`. Takes `{ store, providers, baseUrl }` as explicit deps instead of reaching into portal singleton.
+- **Updated `packages/mcp-core/package.json`** — added `@aomi-labs/client: "workspace:*"` dependency.
+- **Updated `packages/mcp-core/src/index.ts`** — added exports for `buildBackendPort`, `BackendPortDeps`, `buildAuthPort`, `AuthPortDeps`.
+- **Rewrote `apps/portal/src/lib/aomi-auth/mcp-server.ts`** — simplified to ~15 lines: reads portal config, injects into package-level factories from `@aomi-labs/mcp-core`.
+- **Deleted `apps/portal/src/lib/aomi-auth/mcp-backend-bridge.ts`** — replaced by `packages/mcp-core/src/backends/aomi-client.ts`.
+- `auth-config.ts` and `env.ts` remain in portal (deployment-specific singletons).
+- All type checks pass (mcp-core + portal), `build:lib` clean.
 
 ### Release version bumps for publish (2026-04-27)
